# packages/backend/src/OpenApi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "bio": {
            "type": "string"
          },
          "candidate_id": {
            "type": "string"
          },
          "candidate_name": {
            "type": "string"
          },
          "candidate_url": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "party": {
            "type": "string"
          },
          "partyUrl": {
            "type": "string"
          },
          "party_url": {
            "type": "string"
          },
          "photo_filename": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Election": {
        "properties": {
          "admin_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "audit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "auth_key": {
            "type": "string"
          },
          "claim_key_hash": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "credential_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "election_id": {
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "frontend_url": {
            "type": "string"
          },
          "head": {
            "type": "boolean"
          },
          "is_public": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "races": {
            "items": {
              "$ref": "#/components/schemas/Race"
            },
            "type": "array"
          },
          "settings": {
            "$ref": "#/components/schemas/ElectionSettings"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/ElectionState"
          },
          "support_email": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "ElectionRoll": {
        "properties": {
          "address": {
            "type": "string"
          },
          "ballot_id": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "election_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_data": {
            "properties": {
              "inviteResponse": {},
              "reminderResponse": {}
            },
            "type": "object"
          },
          "head": {
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ElectionRollAction"
            },
            "type": "array"
          },
          "ip_hash": {
            "type": "string"
          },
          "precinct": {
            "type": "string"
          },
          "registration": {},
          "state": {
            "$ref": "#/components/schemas/ElectionRollState"
          },
          "submitted": {
            "type": "boolean"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "voter_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ElectionRollAction": {
        "properties": {
          "action_type": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "timestamp": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "ElectionRollState": {
        "enum": [
          "approved",
          "flagged",
          "registered",
          "invalid"
        ],
        "type": "string"
      },
      "ElectionSettings": {
        "properties": {
          "ballot_updates": {
            "type": "boolean"
          },
          "break_ties_randomly": {
            "type": "boolean"
          },
          "invitation": {
            "enum": [
              "address",
              "email"
            ],
            "type": "string"
          },
          "max_rankings": {
            "type": "number"
          },
          "public_results": {
            "type": "boolean"
          },
          "random_candidate_order": {
            "type": "boolean"
          },
          "reminders": {
            "type": "boolean"
          },
          "require_instruction_confirmation": {
            "type": "boolean"
          },
          "term_type": {
            "enum": [
              "election",
              "poll"
            ],
            "type": "string"
          },
          "time_zone": {
            "type": "string"
          },
          "voter_access": {
            "enum": [
              "closed",
              "open",
              "registration"
            ],
            "type": "string"
          },
          "voter_authentication": {
            "$ref": "#/components/schemas/authentication"
          }
        },
        "type": "object"
      },
      "ElectionState": {
        "enum": [
          "archived",
          "closed",
          "draft",
          "finalized",
          "open"
        ],
        "type": "string"
      },
...
      "Race": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "num_winners": {
            "type": "number"
          },
          "precincts": {
            "items": {
              "additionalProperties": false,
              "patternProperties": {
                "^[0-9]+$": {
                  "type": "string"
                }
              },
              "properties": {
                "length": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "race_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "voting_method": {
            "$ref": "#/components/schemas/VotingMethod"
          }
        },
        "type": "object"
      },
...
      "VotingMethod": {
        "enum": [
          "Approval",
          "IRV",
          "Plurality",
          "RankedRobin",
          "STAR",
          "STAR_PR",
          "STV"
        ],
        "type": "string"
      },
...
      "authentication": {
        "properties": {
          "address": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          },
          "ip_address": {
            "type": "boolean"
          },
          "phone": {
            "type": "boolean"
          },
          "registration_api_endpoint": {
            "type": "string"
          },
          "registration_data": {
            "items": [
              {
                "$ref": "#/components/schemas/registration_field"
              }
            ],
            "maxItems": 1,
            "minItems": 1,
            "type": "array"
          },
          "voter_id": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "registration_field": {
        "properties": {
          "field_name": {
            "type": "string"
          },
          "field_type": {
            "enum": [
              "photo",
              "text"
            ],
            "type": "string"
          },
          "help_text": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "paths": {
...
    "/Election/{id}/register": {
      "post": {
        "summary": "Register a voter for an election",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Voter registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "NewElectionRoll": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/rolls": {
      "get": {
        "summary": "Get rolls by election ID",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of rolls",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "electionRollEntry": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      },
      "post": {
        "summary": "Add a new roll to an election",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "NewElectionRoll": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      },
      "put": {
        "summary": "Edit an election roll",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "electionRollEntry": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/rolls/{voter_id}": {
      "get": {
        "summary": "Get roll by voter ID",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          },
          {
            "in": "path",
            "name": "voter_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The voter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "electionRollEntry": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Roll not found"
          }
        }
      }
    },
    "/Election/{id}/rolls/approve": {
      "post": {
        "summary": "Approve an election roll",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll approved"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/rolls/flag": {
      "post": {
        "summary": "Flag an election roll",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll flagged"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/rolls/invalidate": {
      "post": {
        "summary": "Invalidate an election roll",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll invalidated"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/rolls/unflag": {
      "post": {
        "summary": "Unflag an election roll",
        "tags": [
          "Rolls"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roll unflagged"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    }
  },
...
}